Feature/consolidade user login

See merge request cultura-inglesa/ci-service!15

+++ knex/migrations/20200723144778_cycle.ts
@@ -4,18 +4,21 @@ import { setUTF8Table } from "../utils/set-utf8-table.migration";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.createTableIfNotExists(CYCLE_TABLE, (table) => {
-        setUTF8Table(table);
-        table.increments('id');
-        table.string('name', 100);
-        table.boolean('active').notNullable().defaultTo(true);
+    const hasTable = await knex.schema.hasTable(CYCLE_TABLE);
+    if (!hasTable) {
+        await knex.schema.createTable(CYCLE_TABLE, (table) => {
+            setUTF8Table(table);
+            table.increments('id');
+            table.string('name', 100);
+            table.boolean('active').notNullable().defaultTo(true);
 
-        table.integer('levelThemeId').unsigned().notNullable();
-        table.foreign('levelThemeId').references('level_theme.id').onDelete('CASCADE');
+            table.integer('levelThemeId').unsigned().notNullable();
+            table.foreign('levelThemeId').references('level_theme.id').onDelete('CASCADE');
 
-        table.index('active');
-        table.index('levelThemeId');
-    })
+            table.index('active');
+            table.index('levelThemeId');
+        })
+    }
 }
