+++ knex/migrations/20200727205004_user-role_table.ts
@@ -4,14 +4,17 @@ import { setUTF8Table } from "../utils/set-utf8-table.migration";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.createTableIfNotExists(USER_ROLE_TABLE, (table) => {
-        setUTF8Table(table);
-        table.increments('id');
-        table.integer('userId').unsigned().notNullable();
-        table.integer('roleId', 2).unsigned().notNullable();
-        table.foreign('userId').references('user.id').onDelete('CASCADE');
-        table.index('userId');
-    });
+    const hasTable = await knex.schema.hasTable(USER_ROLE_TABLE);
+    if (!hasTable) {
+        await knex.schema.createTable(USER_ROLE_TABLE, (table) => {
+            setUTF8Table(table);
+            table.increments('id');
+            table.integer('userId').unsigned().notNullable();
+            table.integer('roleId', 2).unsigned().notNullable();
+            table.foreign('userId').references('user.id').onDelete('CASCADE');
+            table.index('userId');
+        });
+    }
 }
