+++ knex/migrations/20200817105759_activity_timer_table.ts
@@ -6,20 +6,23 @@ import { CYCLE_ACTIVITY_TABLE } from "../../src/entities/cycle-activity.entity";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.createTable(ACTIVITY_TIMER_TABLE, (table) => {
-        setUTF8Table(table);
-        table.increments('id')
-        table.integer('userId').unsigned().notNullable().references(`${USER_TABLE}.id`).onDelete("CASCADE");
-        table.integer('cycleActivityId').unsigned().notNullable().references(`${CYCLE_ACTIVITY_TABLE}.id`).onDelete("CASCADE");
-        table.dateTime('startTime').notNullable().defaultTo(knex.fn.now());
-        table.dateTime('completionTime').defaultTo(knex.fn.now());
-        table.boolean('completed').defaultTo(false);
+    const hasTable = await knex.schema.hasTable(ACTIVITY_TIMER_TABLE);
+    if (!hasTable) {
+        await knex.schema.createTable(ACTIVITY_TIMER_TABLE, (table) => {
+            setUTF8Table(table);
+            table.increments('id')
+            table.integer('userId').unsigned().notNullable().references(`${USER_TABLE}.id`).onDelete("CASCADE");
+            table.integer('cycleActivityId').unsigned().notNullable().references(`${CYCLE_ACTIVITY_TABLE}.id`).onDelete("CASCADE");
+            table.dateTime('startTime').notNullable().defaultTo(knex.fn.now());
+            table.dateTime('completionTime').defaultTo(knex.fn.now());
+            table.boolean('completed').defaultTo(false);
 
-        table.index('userId');
-        table.index('cycleActivityId');
-        // also indexing the completed field for query performance
-        table.index('completed');
-    });
+            table.index('userId');
+            table.index('cycleActivityId');
+            // also indexing the completed field for query performance
+            table.index('completed');
+        });
+    }
 }
